+++ src/design/plone/contenttypes/interfaces/unita_organizzativa.py
@@ -29,7 +29,7 @@ class IUnitaOrganizzativa(model.Schema, IDesignPloneContentType):
 
     legami_con_altre_strutture = RelationList(
         title=_(
-            "legami_altre_strutture_label", default="Servizi o uffici di riferimento"
+            "legami_altre_strutture_label", default="Strutture o uffici di riferimento"
         ),
         default=[],
         description=_(
@@ -203,7 +203,7 @@ class IUnitaOrganizzativa(model.Schema, IDesignPloneContentType):
     # custom fieldsets and order
     model.fieldset(
         "cosa_fa",
-        label=_("cosa_fa_label", default="Cosa fa"),
+        label=_("cosa_fa_label", default="Competenze"),
         fields=["competenze"],
     )
     model.fieldset(